# specs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: 熊舍管家 MVP API
  version: 1.0.1
  summary: 熊舍管家 MVP 的 REST API 契约
  description: |
    面向 Flutter 客户端的模块化单体 REST API 草案。

    所有租户业务数据都由 Bearer 令牌中的认证上下文确定 owner_id；
    普通写请求体不接收 owner_id。状态迁移统一通过动作接口完成，
    客户端不得直接 PATCH state。

    除公开分享读取外，所有资源 ID 都先在认证 owner_id 范围内解析；
    不存在与跨 owner 资源统一返回 404，错误体不暴露其他账号的业务字段。

    写请求统一支持 Idempotency-Key。可并发编辑的资源通过 If-Match
    传入当前版本，响应同时返回 ETag 与资源 version。所有 date-time
    均以 UTC 传输，业务日期计算使用请求或当前熊舍的 IANA timezone。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.scolvpet.cn/v1
    description: 生产环境
  - url: https://api-staging.scolvpet.cn/v1
    description: 预发布环境
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [任务提醒]
      operationId: listTasks
      summary: 列出任务
      description: 使用 cursor 分页，支持状态、优先级、目标和时间范围筛选。
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskState'
        - name: priority
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskPriority'
        - name: target_type
          in: query
          required: false
          schema:
            type: string
        - name: due_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareTaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [任务提醒]
      operationId: createTask
      summary: 创建手工任务
      description: 系统生成任务也使用同一资源模型；关闭系统通知不影响任务存在。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CareTaskCreateRequest'
      responses:
        '201':
          description: 任务已创建
          headers:
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Key:
              $ref: '#/components/headers/IdempotencyKey'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareTaskResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /tasks/{task_id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [任务提醒]
      operationId: getTask
      summary: 获取任务
      description: 返回阶段进度、关联对象和版本。
      responses:
        '200':
          description: 任务详情
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareTaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [任务提醒]
      operationId: updateTask
      summary: 更新任务非状态字段
      description: 可修改计划时间、优先级、标题和备注；完成状态走 complete 动作。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/CareTaskUpdateRequest'
      responses:
        '200':
          description: 任务已更新
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Key:
              $ref: '#/components/headers/IdempotencyKey'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareTaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /tasks/{task_id}/complete:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [任务提醒]
      operationId: completeTask
      summary: 完成任务或逐项完成任务成员
      description: 支持整窝任务逐只完成；全部 subject 完成或登记例外后任务自动关闭。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskRequest'
            example:
              completed_at: '2026-08-10T02:00:00Z'
              subject_results:
                - subject_id: 018f47a2-73b3-762e-8498-07b13dc3b599
                  status: completed
                  completion_record_id: 018f47a2-951e-7123-a5b3-1ade38a2b49e
                - subject_id: 018f47a2-748c-7d2d-a19f-ab53e616d7d8
                  status: excepted
                  exception_reason: 当日医疗观察，延后称重
              notes: 本次完成 2 项
      responses:
        '200':
          description: 任务进度已更新
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Key:
              $ref: '#/components/headers/IdempotencyKey'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteTaskResponse'
              example:
                data:
                  task:
                    id: 018f47a2-96a7-7e37-a202-cefdc69456ce
                    task_type: pup_weight_check
                    target_type: litter
                    target_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                    scheduled_at: '2026-08-10T00:00:00Z'
                    priority: high
                    state: in_progress
                    stage_total: 5
                    stage_done: 2
                    version: 3
                  item_results:
                    - subject_id: 018f47a2-73b3-762e-8498-07b13dc3b599
                      status: succeeded
                      completion_record_id: 018f47a2-951e-7123-a5b3-1ade38a2b49e
                    - subject_id: 018f47a2-748c-7d2d-a19f-ab53e616d7d8
                      status: succeeded_with_exception
                  auto_closed: false
                meta:
                  request_id: req_01J0TASK001
                  generated_at: '2026-08-10T02:00:01Z'
                  timezone: Asia/Shanghai
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /reminders:
    get:
      tags: [任务提醒]
      operationId: listReminders
      summary: 列出提醒
      description: 返回站内提醒及各通知通道投递状态。
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ReminderState'
        - name: rule_code
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 提醒列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reminders/{reminder_id}:
    parameters:
      - $ref: '#/components/parameters/ReminderId'
    get:
      tags: [任务提醒]
      operationId: getReminder
      summary: 获取提醒
      description: 返回提醒规则、基准事件和多通道投递状态。
      responses:
        '200':
          description: 提醒详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 手机验证码登录后取得的访问令牌
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        写请求唯一键。唯一域为 owner_id + action_code + resource_id + key；相同规范化
        载荷返回首次结果，不同载荷返回 409 IDEMPOTENCY_PAYLOAD_MISMATCH。结果至少保留
        24 小时；confirm-birth、individualize 与分享撤销保留至对应业务记录归档。
      schema:
        type: string
        minLength: 8
        maxLength: 128
      example: 018f47a2-281b-79e2-b861-bf785ab6fba7
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 当前资源版本对应的 ETag，例如双引号包裹的整数版本。
      schema:
        type: string
        pattern: '^W?/"[0-9]+"$'
      example: '"7"'
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: 上一页响应返回的不透明 next_cursor。
      schema:
        type: string
        minLength: 1
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      description: 每页数量。
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReminderId:
      name: reminder_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    IdempotencyKey:
      description: 服务端确认使用的幂等键
      schema:
        type: string
      example: 018f47a2-281b-79e2-b861-bf785ab6fba7
    IdempotencyReplayed:
      description: true 表示本次返回首次请求的已保存结果
      schema:
        type: boolean
      example: false
    ETag:
      description: 当前资源 version 的弱或强 ETag
      schema:
        type: string
        pattern: '^W?/"[0-9]+"$'
      example: '"8"'
    Location:
      description: 新建资源或异步作业的相对地址
      schema:
        type: string
      example: /v1/jobs/018f47a2-7862-7129-b21f-e415c86bbf5c
  responses:
    Unauthorized:
      description: 访问令牌缺失、无效或过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: 请重新登录
              field_errors: []
              recovery_actions:
                - action: login
                  label: 使用手机验证码登录
                  method: POST
                  path: /v1/auth/sessions
            meta:
              request_id: req_01J0ERR401
              generated_at: '2026-07-16T03:00:00Z'
              timezone: Asia/Shanghai
    NotFound:
      description: 资源不存在、已撤销或不属于当前 owner
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 未找到指定资源
              field_errors: []
              recovery_actions: []
            meta:
              request_id: req_01J0ERR404
              generated_at: '2026-07-16T03:00:00Z'
              timezone: Asia/Shanghai
...
    ValidationError:
      description: 字段格式或领域规则校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            field_validation:
              summary: 字段校验错误
              value:
                error:
                  code: VALIDATION_FAILED
                  message: 请求字段校验失败
                  field_errors:
                    - field: initial_alive_count
                      code: minimum
                      message: 初始活仔数必须大于等于 0
                  recovery_actions:
                    - action: edit_request
                      label: 修改后重新提交
                      method: POST
                      path: /v1/breeding-plans/018f47a2-4e63-7cc4-80b4-94a589917a36/confirm-birth
                meta:
                  request_id: req_01J0ERR422A
                  generated_at: '2026-07-16T03:00:00Z'
                  timezone: Asia/Shanghai
            business_rule:
              summary: 领域完整性错误
              value:
                error:
                  code: LITTER_COUNT_MISMATCH
                  message: 窝仔数量未闭合
                  field_errors: []
                  current_version: 5
                  recovery_actions:
                    - action: inspect_reconciliation
                      label: 查看数量对账
                      method: GET
                      path: /v1/litters/018f47a2-72b3-76d8-b31f-5848c42bb577
                    - action: adjust_count
                      label: 补录数量流水
                      method: POST
                      path: /v1/litters/018f47a2-72b3-76d8-b31f-5848c42bb577/count-events
                  details:
                    expected_managed_count: 5
                    represented_count: 4
                    difference: 1
                meta:
                  request_id: req_01J0ERR422B
                  generated_at: '2026-07-16T03:00:00Z'
                  timezone: Asia/Shanghai
...
  schemas:
...
    TaskState:
      type: string
      enum: [pending, in_progress, completed, snoozed, cancelled, superseded]
    TaskPriority:
      type: string
      enum: [low, normal, high, urgent, critical]
    ReminderState:
      type: string
      enum: [pending, sent, read, failed, superseded, cancelled]
...
    CareTaskCreateRequest:
      type: object
      required: [task_type, target_type, target_id, scheduled_at, priority]
      properties:
        task_type:
          type: string
        target_type:
          type: string
        target_id:
          type: string
          format: uuid
        title:
          type: [string, 'null']
          maxLength: 200
        scheduled_at:
          type: string
          format: date-time
        priority:
          $ref: '#/components/schemas/TaskPriority'
        subject_ids:
          type: array
          items:
            type: string
            format: uuid
        notes:
          type: [string, 'null']
          maxLength: 4000
      additionalProperties: false
    CareTaskUpdateRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: [string, 'null']
          maxLength: 200
        scheduled_at:
          type: string
          format: date-time
        priority:
          $ref: '#/components/schemas/TaskPriority'
        notes:
          type: [string, 'null']
          maxLength: 4000
      additionalProperties: false
    CareTaskResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/CareTask'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      additionalProperties: false
    CareTaskListResponse:
      type: object
      required: [data, page, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CareTask'
        page:
          $ref: '#/components/schemas/PageInfo'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      additionalProperties: false
    CompleteTaskRequest:
      type: object
      required: [completed_at, subject_results]
      properties:
        completed_at:
          type: string
          format: date-time
        subject_results:
          type: array
          minItems: 1
          items:
            type: object
            required: [subject_id, status]
            properties:
              subject_id:
                type: string
                format: uuid
              status:
                type: string
                enum: [completed, excepted]
              completion_record_id:
                type: [string, 'null']
                format: uuid
              exception_reason:
                type: [string, 'null']
                maxLength: 1000
            additionalProperties: false
        notes:
          type: [string, 'null']
          maxLength: 2000
      additionalProperties: false
...
    ReminderListResponse:
      type: object
      required: [data, page, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Reminder'
        page:
          $ref: '#/components/schemas/PageInfo'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      additionalProperties: false
...